requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
distribute==0.6.27
python-dateutil==2.1
redis==2.4.13
six==1.1.0
twitter==1.8.0
wsgiref==0.1.2
